// schedule.web.application/src/common/enums/HttpMethod.ts
export enum HttpMethod {
    GET = 'GET',
    POST = 'POST',
    PUT = 'PUT',
    DELETE = 'DELETE',
    HEAD = 'HEAD',
    CONNECT = 'CONNECT',
    OPTIONS = 'OPTIONS',
    TRACE = 'TRACE',
    PATCH = 'PATCH',
    PROPFIND = 'PROPFIND',
    PROPPATCH = 'PROPPATCH',
    MKCOL = 'MKCOL',
    COPY = 'COPY',
    MOVE = 'MOVE',
    LOCK = 'LOCK',
    UNLOCK = 'UNLOCK',
}